Add the basics of a database
This uses aiopg and sqlalchemy. Right now we only have one table,
replays. I'm sure the fields in that table will be changed.

Right now we have a basics 'add replay to table' and 'list all replays',
which is a decent start.

Also included a script `create_db.py` that will use a non-async
sqlalchemy to create the tables. See caveats in the comments of that
file.

This final PR doesn't look much like it, but this change grew out of
https://gist.github.com/samuelcolvin/e471e3e5869b13ef0669