Make heatmaps from the maps/<uuid>/replay endpoint
This uses the de_heatmp js from the previous commit. It also puts some
helpful links to the heatmaps in the /maps template so they are easy to
click.

It suffers from being a 3-4 second load on my machine, I'm not sure
exactly why consideing there isn't much data. Possibly this is due to
the app running on my laptop and the database being in the cloud.