Encode at the edges
This encodes the temp_url_key to bytes in the config reading step, so no
other parts of the application ever have to worry about the encoding.

This also encodes strings liberally in the create_tempurl function
because b"\n".join(stuff) requires the stuff to be bytes, it doesn't
automatically encode strings.